api: Make the path builder an associated type and constant
This will allow with to use new types to stop requesting the
`SupportedVersions` on endpoints that cannot have them or don't use
them.
